fix(studio): guard table move on header overlap, restore modal focus trap
- Hide Move row/column when the selection overlaps a header axis, matching
  the existing Delete guards so header-typed cells cannot be displaced.
- Keep TableBubbleMenu hidden while a Chakra modal dialog is open so
  FocusLock no longer fights the BubbleMenu tab surface; re-enable the
  default focus trap on TableSettingsModal.
- Move TableBubbleMenu tests under __tests__/ per features conventions.
