+++ apps/studio/src/components/PageEditor/TableSettingsModal.tsx
@@ -68,11 +68,7 @@ export const TableSettingsModal = ({
   }, [isOpen])
 
   return (
-    // trapFocus=false: with TableBubbleMenu mounted, Chakra's FocusLock and
-    // TipTap's BubbleMenu (tabIndex=0 + blur/focus handlers) fight when the
-    // editor blurs into this modal — hanging the tab. Keep autoFocus so the
-    // textarea still receives focus; just don't run the focus trap.
-    <Modal isOpen={isOpen} onClose={onClose} trapFocus={false}>
+    <Modal isOpen={isOpen} onClose={onClose}>
       <ModalOverlay />
 
       <ModalContent>
