+++ apps/studio/src/features/editing-experience/components/TableBubbleMenu/__tests__/TableBubbleMenu.browser.test.tsx
@@ -10,7 +10,7 @@ import { afterEach, describe, expect, it } from "vitest"
 import { useTextEditor } from "~/features/editing-experience/hooks/useTextEditor"
 import { theme } from "~/theme"
 
-import { TableBubbleMenu } from "./TableBubbleMenu"
+import { TableBubbleMenu } from "../TableBubbleMenu"
 
 const SEED_CONTENT: JSONContent = {
   type: "prose",
@@ -256,11 +256,11 @@ describe("TableBubbleMenu", () => {
     expect(queryByText("Move down")).toBeNull()
     expect(await findByText("Move up")).toBeTruthy()
 
-    // Header row (top)
+    // Header row (top) — move is withheld when the selection overlaps the header.
     selectCells(editor, 0, 2)
     await activateTableBubbleMenu(findByRole)
     expect(queryByText("Move up")).toBeNull()
-    expect(await findByText("Move down")).toBeTruthy()
+    expect(queryByText("Move down")).toBeNull()
   })
 
   it("moves a multi-column selection as a block (A,B → right becomes C,A,B)", async () => {
@@ -279,7 +279,7 @@ describe("TableBubbleMenu", () => {
     expect(firstRowTexts(editor)).toEqual(["Column C", "Column A", "Column B"])
   })
 
-  it("excludes Delete row when the selection includes the header row", async () => {
+  it("excludes Delete row and Move up/down when the selection includes the header row", async () => {
     const { editor, findByText, findByRole, queryByText, queryByRole } =
       await renderHarness()
 
@@ -290,21 +290,24 @@ describe("TableBubbleMenu", () => {
     expect(headerToggle).toBeChecked()
     expect(await findByText("Add row above")).toBeTruthy()
     expect(await findByText("Add row below")).toBeTruthy()
-    expect(await findByText("Move down")).toBeTruthy()
-    // Header axes withhold Delete — unset the header first, then delete as a
-    // body row/column — including when the selection also spans body rows.
+    // Header axes withhold Delete and Move — unset the header first, then
+    // delete or reorder as a body row/column — including when the selection
+    // also spans body rows.
     expect(queryByText("Delete row")).toBeNull()
     expect(queryByText("Move up")).toBeNull()
+    expect(queryByText("Move down")).toBeNull()
 
-    // Header row + first body row: still withhold Delete, but the Header
-    // switch is only for the exact top row.
+    // Header row + first body row: still withhold Delete and Move, but the
+    // Header switch is only for the exact top row.
     selectCells(editor, 0, 5)
     await activateTableBubbleMenu(findByRole)
     expect(queryByRole("checkbox", { name: "Header row" })).toBeNull()
     expect(queryByText("Delete row")).toBeNull()
+    expect(queryByText("Move up")).toBeNull()
+    expect(queryByText("Move down")).toBeNull()
   })
 
-  it("excludes Delete column when the selection includes a header column", async () => {
+  it("excludes Delete column and Move left/right when the selection includes a header column", async () => {
     const { editor, findByText, findByRole, queryByText, queryByRole } =
       await renderHarness()
 
@@ -315,6 +318,7 @@ describe("TableBubbleMenu", () => {
       await findByRole("checkbox", { name: "Header column" }),
     ).not.toBeChecked()
     expect(await findByText("Delete column")).toBeTruthy()
+    expect(await findByText("Move right")).toBeTruthy()
 
     act(() => {
       editor.chain().focus().toggleHeaderColumn().run()
@@ -326,13 +330,17 @@ describe("TableBubbleMenu", () => {
       await findByRole("checkbox", { name: "Header column" }),
     ).toBeChecked()
     expect(queryByText("Delete column")).toBeNull()
+    expect(queryByText("Move left")).toBeNull()
+    expect(queryByText("Move right")).toBeNull()
 
-    // Header column + next column: still withhold Delete; Header switch only
-    // for the exact leftmost column.
+    // Header column + next column: still withhold Delete and Move; Header
+    // switch only for the exact leftmost column.
     selectCells(editor, 0, 7)
     await activateTableBubbleMenu(findByRole)
     expect(queryByRole("checkbox", { name: "Header column" })).toBeNull()
     expect(queryByText("Delete column")).toBeNull()
+    expect(queryByText("Move left")).toBeNull()
+    expect(queryByText("Move right")).toBeNull()
   })
 
   it("shows only Merge cells for an irregular multi-cell selection", async () => {
